+++ source/renderer/app/config/sidebarConfig.js
@@ -2,6 +2,7 @@
 import { ROUTES } from '../routes-config';
 import walletsIcon from '../assets/images/sidebar/wallet-ic.inline.svg';
 import hardwareWalletsIcon from '../assets/images/sidebar/hardware-wallet-ic.inline.svg';
+import redeemItnRewardsIcon from '../assets/images/sidebar/redeem-itn-rewards-ic.inline.svg';
 import settingsIcon from '../assets/images/sidebar/settings-ic.inline.svg';
 import paperWalletCertificateIcon from '../assets/images/sidebar/paper-certificate-ic.inline.svg';
 import delegationIcon from '../assets/images/sidebar/delegation-ic.inline.svg';
@@ -24,6 +25,11 @@ export const CATEGORIES_BY_NAME = {
     name: 'HARDWARE_WALLETS',
     icon: hardwareWalletsIcon,
     route: ROUTES.HARDWARE_WALLETS.ROOT,
+  },
+  REDEEM_ITN_REWARDS: {
+    name: 'REDEEM_ITN_REWARDS',
+    icon: redeemItnRewardsIcon,
+    route: ROUTES.REDEEM_ITN_REWARDS,
   },
   PAPER_WALLET_CREATE_CERTIFICATE: {
     name: 'PAPER_WALLET_CREATE_CERTIFICATE',
@@ -54,13 +60,14 @@ export const CATEGORIES_BY_NAME = {
 
 export const CATEGORIES = [
   CATEGORIES_BY_NAME.WALLETS,
-  // CATEGORIES_BY_NAME.PAPER_WALLET_CREATE_CERTIFICATE,
+  CATEGORIES_BY_NAME.REDEEM_ITN_REWARDS,
   CATEGORIES_BY_NAME.SETTINGS,
   CATEGORIES_BY_NAME.NETWORK_INFO,
 ];
 
 export const CATEGORIES_WITHOUT_NETWORK_INFO = [
   CATEGORIES_BY_NAME.WALLETS,
+  CATEGORIES_BY_NAME.REDEEM_ITN_REWARDS,
   CATEGORIES_BY_NAME.SETTINGS,
 ];