+++ source/renderer/app/config/stakingConfig.js
@@ -1,5 +1,7 @@
 // @flow
 
+import type { RedeemItnRewardsStep } from '../types/stakingTypes';
+
 export const MIN_DELEGATION_FUNDS = 10;
 
 // Dimensions
@@ -27,3 +29,14 @@ export const STAKE_POOL_TRANSACTION_CHECK_INTERVAL = 1 * 1000; // 1 second | uni
 export const STAKE_POOL_TRANSACTION_CHECKER_TIMEOUT = 30 * 1000; // 30 seconds | unit: milliseconds;
 export const STAKE_POOLS_INTERVAL = 1 * 60 * 1000; // 1 minute | unit: milliseconds;
 export const STAKE_POOLS_FAST_INTERVAL = 1 * 1000; // 1 second | unit: milliseconds;
+
+// Redeem ITN Rewards
+
+export const REDEEM_ITN_REWARDS_STEPS: {
+  [key: string]: RedeemItnRewardsStep,
+} = {
+  CONFIGURATION: 'configuration',
+  CONFIRMATION: 'confirmation',
+  SUCCESS: 'success',
+  FAILURE: 'failure',
+};