+++ source/renderer/app/types/stakingTypes.js
@@ -1,0 +1,7 @@
+// @flow
+
+export type RedeemItnRewardsStep =
+  | 'configuration'
+  | 'confirmation'
+  | 'success'
+  | 'failure';